Account Number Setter
Added setter function to allow independent setting of Account Number. This is meant to be used for attempted loading of an account as the load has a dependency that requires the setting of Account Number.
Throws a NumOutOfBounds error if value is not between 1 and 99999999.
Added an Account constructor for when only the Account Number is known. This will then attempt to load the account file whose 8-digit ID matches.
Throws a NumOutOfBounds error if value is not between 1 and 99999999.